ip2gid: Move client code to a separate file
Move client code to a separate file client.c.
